requirements.txt
streamlit>=1.28.0
streamlit-webrtc>=0.47.0
opencv-python-headless>=4.8.0
ultralytics>=8.0.0
numpy
googlemaps
python-dotenv